# swagger.yaml
openapi: 3.0.3
info:
  title: Aether AI Gateway
  version: 1.0.0
  description: 'HTTP API for user, session, and improvement plan report workflows in the Aether core gateway. Every response carries an X-Request-Id header: the identifier of that request in the hub_metrics base.'
servers:
  - url: /
...
paths:
  /aether-api/v1/ai/user/{id_external_user}:
    get:
      tags:
        - Users
      summary: Get user by external identifier
      description: Retrieves a user profile using the external user identifier stored in the core database.
      operationId: getUserByExternalId
      parameters:
        - name: id_external_user
          in: path
          required: true
          description: External user identifier.
          schema:
            type: integer
          example: 12345
      responses:
        '200':
          description: User profile found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseData'
              example:
                id_external_user: 12345
                role: analyst
                usecase: report_generation
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: User with id_external_user 12345 not found.
        '503':
          description: Database connection is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Database is not initialized
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: 'Error retrieving user: Error fetching user from database: ...'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
    UserResponseData:
      type: object
      properties:
        id_external_user:
          type: integer
          description: External identifier of the user.
          example: 12345
        role:
          type: string
          description: Role assigned to the user.
          example: analyst
        usecase:
          type: string
          description: User use case or profile category.
          example: report_generation
      required:
        - id_external_user
        - role
        - usecase
...